requirements.txt
# Django and Core Dependencies
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0

# Database
mysqlclient==2.2.0

# Face Recognition and Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3

# Async Task Processing
celery==5.3.4
redis==5.0.1

# Email and Notifications
django-email-backends==0.1.0

# Security
cryptography==41.0.7
bcrypt==4.1.2

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2

# Optional: Enhanced face recognition (uncomment if needed)
# deepface==0.0.79
# tensorflow==2.13.0
# torch==2.1.0
# torchvision==0.16.0

# Optional: SilentFace anti-spoofing (uncomment if needed)
# silentface==1.0.0
